requirements.txt
flask
supabase-py
gunicorn
python-dotenv
psycopg2